FEAT: All warnings generated by the Fiftyon.Common.Tests are now treated as error when built with the provided solution. The warning level is set at level 4 /W4. Generally, /W3 is recommended for production quality. Setting to /W4 just to ensure the code can be clean as possible. /Wall will generate too many unuseful warnings and unnecessary so not recommended. - Some of the warnings can only be fixed by suppression such as warnings with different implementations of a function prototype, where not all parameters are required to be used (e.g. Mock implementation in tests.) - Warnings that are generated by the third party google gtest.h file are allso supressed as not within our control.

+++ ComponentMetaData.cpp
@@ -26,7 +26,7 @@ using namespace FiftyoneDegrees::Common;
 
 ComponentMetaData::ComponentMetaData()
     : ComponentMetaData(
-        -1, // component id
+        (byte)-1, // component id
         "invalid", // name
         0 // default profile id
     ) { }
